==> 1.1 Data Gathering.ipynb <==
1.1 Parse Ulta site list of Skincare and Makeup products

import re
import time
import random
import warnings
import pandas as pd
import numpy as np
from typing import List
from lxml import etree as et
from bs4 import BeautifulSoup
from time import sleep

#pip install ediblepickle
#pip install selenium
#pip install webdriver-manager

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.chrome.options import Options
from selenium.common.exceptions import NoSuchElementException

chrome_options=Options()
chrome_options.add_argument('--no-sandbox')   
chrome_options.add_argument('--disable-dev-shm-usage')
chrome_options.add_argument('--disable-gpu')

warnings.filterwarnings('ignore')
#driver = webdriver.Chrome(ChromeDriverManager().install())
driver = webdriver.Chrome(chrome_options=chrome_options)

#Handle the exception of product without Star Ratings
def get_attr(x,tag,class_name):
    try:
        return x.find(tag,class_name).text
    except Exception:
        return None

#Parse Brand,ProductID,ProductName,Price,Rating from all results page
#Fetch link for each product

def get_product_basic(url):
    i=1
    prod_basic=[]
    while (True):
        page=url+"?page={}".format(i)
        driver.get(page)
        sleep(4)
        soup = BeautifulSoup(driver.page_source, 'html.parser')
        prod_listing=soup.find('ul',class_="ProductListingResults__productList ProductListingResults__productList--space--top")
        try:
            prod=prod_listing.find_all('li',class_='ProductListingResults__productCard')    
        except Exception:
            return prod_basic
        
        each_page=[ {"Brand":x.find('span','ProductCard__brand').text,
                     "Name":x.find('span','ProductCard__product').text,
                     "Rating":get_attr(x,'div','ProductCard__rating'),
                     "Price":x.find('div','ProductCard__price').text,  
                     "Link":x.find('a')['href'],
                     "ID":x.find('div','ProductCard')['data-item-id']
                    } for x in prod]
        prod_basic.extend(each_page)
        i=i+1
        if(len(prod)<96): break
    return prod_basic

#Parse list of results page for skincare
skincare_url='https://www.ulta.com/shop/skin-care/all'
all_skincare = get_product_basic(skincare_url)
#print("There are {} skincare products in total.".format(len(all_skincare)))
all_skincare[0]

df_1=pd.DataFrame.from_dict(all_skincare)
df_1.head()

df_1.to_csv('Skincare.csv',index=None)

%%time
#Do the same process for skincare
makeup_url="https://www.ulta.com/shop/makeup/all"
all_makeup = get_product_basic(makeup_url)
print("There are {} makeup products in total.".format(len(all_makeup)))
df_2=pd.DataFrame.from_dict(all_makeup)
df_2.to_csv('Makeup.csv',index=None)
df_2.info()

df_2.tail(5)

driver.quit()
