==> 2. Data Cleaning.ipynb <==
import pandas as pd

skincare=pd.read_csv('SkinCareDetail.csv')
makeup=pd.read_csv('MakeupDetail.csv')
skincare.shape,makeup.shape

#Combine makeup and skincare data
skincare.insert(8,'Main_Category',["Skincare"]*4397,True)
makeup.insert(8,'Main_Category',["Makeup"]*4896,True)
ulta=pd.concat([skincare,makeup],axis=0)
ulta.index=range(0,len(ulta))
ulta['Price']=[x.split("Original Price")[0] for x in ulta['Price']]
ulta.sample(3)

indexAge=ulta[ ulta['Category'].isin(['Makeup Brushes & Tools' ,'Makeup Bags & Organizers','Skincare Tools',
                                     'Bath & Body Accessories'
                                     ])].index
ulta.drop(indexAge,inplace=True)

def remove_prefix(text, prefix):
    return text[text.startswith(prefix) and len(prefix):].strip()

ulta['Ingredient']=ulta.apply(lambda x: remove_prefix(x['Ingredient'],'Ingredients\n'),axis=1)
ulta['Ingredient']=ulta.apply(lambda x: remove_prefix(x['Ingredient'],'Active:'),axis=1)
ulta['Ingredient']=ulta.apply(lambda x: x['Ingredient'].lower(),axis=1)

import re
ig=ulta['Ingredient'][1951]
print(ig)
def get_ig_list(ig):
    l=re.split('[,]',ig.split(".")[0])
    l=[x.strip().lower().rstrip(".*") for x in l  ]
    l=[re.sub("\d*%","",x).strip() for x in l if not x.startswith(('*',"please","water","aqua","eau","ingredient")) 
       and not x.isdigit() and "may contain" not in x ]
    return sorted(l)

get_ig_list(ig)

#Drop lines with <=3 ingredients provided
ulta['Ig_list']=ulta.apply(lambda x: get_ig_list(x['Ingredient']),axis=1)
ulta['len']=ulta.apply(lambda x:len(x['Ig_list']),axis=1)
#ulta['Ig_list'][1951]
indexAge=ulta[ ulta['len']<=3].index
ulta.drop(indexAge,inplace=True)
#ulta = ulta.drop(columns=['Ig_list', 'len'])

#Drop duplicates
ulta.drop_duplicates(subset=["Name",'ID'],inplace=True)
ulta.to_csv('Clean_data',index=None)


from platform import python_version

print(python_version())
