==> 4. Recommendation.ipynb <==
# Identify and examine labels to find safe-for-you beauty products

import pandas as pd
import numpy as np
from sklearn import base
from sklearn.feature_extraction import DictVectorizer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.pipeline import Pipeline,FeatureUnion
from sklearn.neighbors import NearestNeighbors
import re
import spacy
from sklearn.decomposition import TruncatedSVD
from spacy.lang.en.stop_words import STOP_WORDS

#1.user input a skincare product
#2.user identify an allergen/irritation
#3.recommend products

#conda install -c conda-forge spacy
#python -m spacy download en
#pip install streamlit-extras

ulta=pd.read_csv('Clean_data')
ulta.sample(3)

product=ulta['Name'][2403]
ig=ulta['Ingredient'][2403]
product,ig

allergen='il58a'

#Check if Ingredient list contains allergen item

def containsAllergen(df,allergen):
    return df[df['Ingredient'].apply(lambda x: allergen not in x)]

## Model Selection:

### Model 1: CountVectorizer, NearestNeighbors

%%time
#CountVectorizer,NearestNeighbors
#based on ingredients of previously working-well products
#check products for problematic ingredients

def recommender_1(df,product, allergen):
    safe_df= df.copy()
    
    bag_of_words_vectorizer=CountVectorizer(min_df=0,
                             ngram_range=(1,2), 
                             stop_words='english')
    counts=bag_of_words_vectorizer.fit_transform(safe_df['Ingredient'])
    nn = NearestNeighbors(n_neighbors=20).fit(counts) 
    index=safe_df[safe_df['Name']==product].index
    #safe_df.index[safe_df['Name']==product][0]
    dists, indices = nn.kneighbors(counts[index[0]])
    prod_nbr=safe_df.iloc[indices[0]]
    
    prod_nbr=prod_nbr[prod_nbr['Main_Category']==df['Main_Category'][index[0]]]
    return containsAllergen(prod_nbr,allergen).head(5)
    #return #.reset_index(drop=True)

prod_1=recommender_1(ulta,product,allergen)
prod_1

### Model 2: CountVectorizer, Cosine similarity

%%time
#CountVectorizer,cosine similarity
"""Calculate similarity scores: Use a content-based filtering algorithm to calculate similarity scores 
between products based on their ingredient lists."""

def cos_similarity(row1, row2):
    dot = np.dot(row1, row2.T)[0, 0]
    len1 = np.sqrt(np.dot(row1, row1.T))[0, 0]
    len2 = np.sqrt(np.dot(row2, row2.T))[0, 0]
    return dot / len1 / len2


def recommender_2(df,product, allergen):
    bag_of_words_vectorizer=CountVectorizer(min_df=0,
                             ngram_range=(1,2), 
                             stop_words='english')
    counts=bag_of_words_vectorizer.fit_transform(df['Ingredient'])
    #print(counts.shape)
    prod_nbr={}
    for j in range(len(counts.toarray())):
        prod_nbr[j]=cos_similarity(counts[2147],counts[j])
    nbr=sorted(prod_nbr.items(),key=lambda x:x[1],reverse=True )
    indices=[x[0] for x in nbr[0:30]]

    prod_nbr=df.iloc[indices] 
    index=df.index[df['Name']==product][0]
    prod_nbr=prod_nbr[prod_nbr['Main_Category']==df['Main_Category'][index]]
    return containsAllergen(prod_nbr,allergen).head(5)

prod_2=recommender_2(ulta,product,allergen)
prod_2

### Model 3: DictVectorizer, NearestNeighbors

%%time
#Feature engineering
class DictEncoder(base.BaseEstimator, base.TransformerMixin):
    
    def __init__(self, col):
        self.col = col
    
    def fit(self, X, y=None):
        return self
    
    def transform(self, X):
        
        def to_dict(l):
            try:
                return {x.strip(): 1 for x in l}
            except TypeError:
                return {}        
        return X[self.col].apply(to_dict)


def recommender_3(df,product, allergen):
    df_1=df.copy()
    #df_1['Ig_list']=df.apply(lambda x:x['Ingredient'].split(','),axis=1)
    ig_pipe = Pipeline([('encoder', DictEncoder('Ig_list')),
                     ('vectorizer', DictVectorizer())#,('svd', TruncatedSVD(n_components=300))
                   ])
    #dt_pipe=Pipeline([("encoder",DictEncoder('Dt_list')),('vectorizer',DictVectorizer())])
    #union=FeatureUnion([('Ingredient',ig_pipe),('Detail',dt_pipe)])
    features = ig_pipe.fit_transform(df_1)
    nn = NearestNeighbors(n_neighbors=20).fit(features)
    index=df_1.index[df_1['Name']==product][0]
    dists, indices = nn.kneighbors(features[index])
    
    prod_nbr=df_1.iloc[indices[0]]    
    prod_nbr=prod_nbr[prod_nbr['Main_Category']==df_1['Main_Category'][index]]
    
   
    return containsAllergen(prod_nbr,allergen).head(5)

prod_3=recommender_3(ulta,product,allergen)
prod_3

prod_1.index,prod_2.index,prod_3.index

prod_1['Name'],prod_2['Name'],prod_3['Name']

#### Model 1 generates the best result with the fastest speed.
